// server/src/utils/types/calculate-burnout.types.ts
export type Metrics = {
  countMessage: number;
  countCharInMessage: number;
  countCommits: number;
  countChangedCode: number;
  countHoursInApps: number;
  countTasks: number;
  countDaysOnTask: number;
  lastUpGradeDate: number;
  countWorkYearInCompany: number;
  countCall: number;
  callDuration: number;
  countWorkHours: number;
  countWorkHoursOver: number;
  countDaysSickTime: number;
  countMissDeadline: number;
  countOutstandingTask: number;
};